refactor: update package.json scripts and dependencies; enhance NodeGraphEditor with new UI primitives and Convex integration
