+++ package.json
@@ -5,7 +5,8 @@
 	"type": "module",
 	"scripts": {
 		"preinstall": "node -e \"if (!process.env.CI && process.env.npm_execpath && !process.env.npm_execpath.includes('pnpm')) { console.error('Please use pnpm (npm i -g pnpm)'); process.exit(1) }\"",
-		"dev": "vite",
+		"dev": "concurrently -c auto -n convex,vite \"pnpm convex:dev\" \"vite\"",
+		"vite:dev": "vite",
 		"build": "tsc -b && vite build",
 		"lint": "eslint .",
 		"preview": "vite preview",
@@ -15,9 +16,12 @@
 		"docs:preview": "vitepress preview docs",
 		"docs:check": "markdownlint \"**/*.md\" && linkinator docs --recurse",
 		"api:docs": "echo 'typedoc disabled in CI for now'",
-		"generate:adapters": "tsx scripts/generate-drei-registry.ts"
+		"generate:adapters": "tsx scripts/generate-drei-registry.ts",
+		"convex:dev": "convex dev --tail-logs",
+		"convex:codegen": "convex codegen"
 	},
 	"dependencies": {
+		"@convex-dev/eslint-plugin": "0.0.1-alpha.4",
 		"@dnd-kit/core": "^6.3.1",
 		"@radix-ui/react-checkbox": "^1.3.2",
 		"@radix-ui/react-context-menu": "^2.2.15",
@@ -42,6 +46,8 @@
 		"class-variance-authority": "^0.7.1",
 		"clsx": "^2.1.1",
 		"cmdk": "^1.1.1",
+		"convex": "^1.16.2",
+		"convex-helpers": "^0.1.104",
 		"d3": "^7.9.0",
 		"lucide-react": "^0.536.0",
 		"motion": "^12.23.12",
@@ -54,6 +60,7 @@
 		"zod": "^3.23.8"
 	},
 	"devDependencies": {
+		"concurrently": "^9.1.2",
 		"@codemirror/lang-javascript": "^6.2.4",
 		"@eslint/js": "^9.32.0",
 		"@playwright/test": "^1.47.2",
@@ -64,6 +71,7 @@
 		"@uiw/react-codemirror": "^4.25.1",
 		"@vitejs/plugin-react": "^4.7.0",
 		"eslint": "^9.32.0",
+		"eslint-plugin-react-compiler": "19.1.0-rc.2",
 		"eslint-plugin-react-hooks": "^5.2.0",
 		"eslint-plugin-react-refresh": "^0.4.20",
 		"globals": "^16.3.0",
